# src/swagger-config.yml
openapi: 3.0.0
info:
  title: My API
  version: '1.0.0'
  description: Translate API
...
security:
  - apiKey: [ ]
paths:
  /translate:
    post:
      tags:
        - Translate
      summary: Translate a text
      description: Translate a text
      operationId: addTranslate
      security:
        - apiKey: [ ]
      requestBody:
        $ref: '#/components/requestBodies/TranslateCreation'
      responses:
        '201':
          $ref: '#/components/responses/Created'
        '400':
          $ref: '#/components/responses/MissingFieldsError'
        '500':
          $ref: '#/components/responses/InternalServerError'
    get:
      tags:
        - Translate
      summary: Get all traductions
      description: Returns all traductions
      operationId: getTraductions
      security:
        - apiKey: [ ]
      responses:
        '200':
          description: Traductions fetched successfully
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/TranslateList'
        '500':
          $ref: '#/components/responses/InternalServerError'
...
  /auth:
    get:
      tags:
          - User
      summary: Get all users
      description: Returns all users
      operationId: getUsers
      security:
        - apiKey: [ ]
      responses:
        '200':
          description: Users fetched successfully
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/UserList'
        '403':
          $ref: '#/components/responses/ForbiddenError'
        '500':
          $ref: '#/components/responses/InternalServerError'
...
components:
  securitySchemes:
    apiKey:
      type: apiKey
      name: Authorization
      in: header
...
  schemas:
    TranslateList:
      type: object
      required: [ text, language, discordId ]
      properties:
        success:
          type: boolean
          default: true
          example: true
        data:
          type: array
          items:
            $ref: '#/components/schemas/Translate'
    Translate:
      type: object
      required: [ text, language, discordId ]
      properties:
        text:
          type: string
          example: "Bonjour à tous"
        language:
          type: string
          example: "en-US"
        discordId:
          type: string
          example: "123456789"
    UserList:
      type: object
      required: [ username, password ]
      properties:
        success:
          type: boolean
          default: true
          example: true
        data:
          type: array
          items:
            $ref: '#/components/schemas/User'
    User:
      type: object
      required: [ username, password ]
      properties:
        username:
          type: string
          example: "username"
        password:
          type: string
          example: "password"
  requestBodies:
    TranslateCreation:
      required: true
      description: Translate creation
      content:
        application/json:
          schema:
            required: [ text, language, discordId ]
            allOf:
              - $ref: '#/components/schemas/Translate'
...
  responses:
...
    Created:
      description: Created
      content:
        application/json:
          schema:
            type: object
            required: [ success, id ]
            properties:
              success:
                type: boolean
                default: true
                example: true
              id:
                type: string
                example: "3fa85f64-5717-4562-b3fc-2c963f66afa6"
...
    MissingFieldsError:
      description: Missing fields
      content:
        application/json:
          schema:
            type: object
            required: [ success, message ]
            properties:
              success:
                type: boolean
                default: false
                example: false
              message:
                type: string
                example: "Missing fields"
    ForbiddenError:
      description: Forbidden
      content:
        application/json:
          schema:
            type: object
            required: [ success, message ]
            properties:
              success:
                  type: boolean
                  default: false
                  example: false
              message:
                  type: string
                  example: "Forbidden"
...
    InternalServerError:
      description: Internal server error
      content:
        application/json:
          schema:
            type: object
            required: [ success, message ]
            properties:
              success:
                type: boolean
                default: false
                example: false
              message:
                type: string
                example: "Internal server error"